+++ sdk/servicebus/service-bus/test/auth.spec.ts
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ServiceBusConnectionStringModel, SharedKeyCredential } from "@azure/core-amqp";
+import { SharedKeyCredential } from "@azure/core-amqp";
 import chai from "chai";
-import { parseConnectionString } from "rhea-promise";
-import { ServiceBusReceiver } from "../src/receivers/receiver";
-import { ServiceBusClient } from "../src/serviceBusClient";
+import { ServiceBusClient, ServiceBusReceiver, parseServiceBusConnectionString } from "../src";
 import { getEnvVars } from "./utils/envVarUtils";
 import { TestClientType } from "./utils/testUtils";
 import {
@@ -32,14 +30,12 @@ type UnpackReturnType<T extends (...args: any) => any> = ReturnType<T> extends P
 
         const { SERVICEBUS_CONNECTION_STRING: serviceBusConnectionString } = getEnvVars();
 
-        const { Endpoint: fqdn } = parseConnectionString<ServiceBusConnectionStringModel>(
-          serviceBusConnectionString
-        );
+        const endpoint = parseServiceBusConnectionString(serviceBusConnectionString).endpoint;
 
         sasConnectionString = getSasConnectionString(
           serviceBusConnectionString,
           entities.queue ?? `${entities.topic!}`,
-          fqdn.replace(/\/+$/, "")
+          endpoint.replace(/\/+$/, "")
         );
       });