+++ sdk/servicebus/service-bus/test/internal/shared.spec.ts
@@ -5,6 +5,7 @@ import { getMessageIterator, wrapProcessErrorHandler } from "../../src/receivers
 import chai from "chai";
 import { ServiceBusReceiver } from "../../src/receivers/receiver";
 import { ServiceBusLogger } from "../../src/log";
+import { ProcessErrorArgs } from "../../src/models";
 const assert = chai.assert;
 
 describe("shared", () => {
@@ -13,8 +14,22 @@ describe("shared", () => {
 
     const wrappedProcessError = wrapProcessErrorHandler(
       {
-        processError: (err) => {
-          assert.equal(err.message, "Actual error that was passed in from service bus to the user");
+        processError: (args: ProcessErrorArgs) => {
+          assert.deepEqual(
+            {
+              message: args.error.message,
+              fullyQualifiedNamespace: args.fullyQualifiedNamespace,
+              entityPath: args.entityPath,
+              errorSource: args.errorSource
+            },
+            {
+              message: "Actual error that was passed in from service bus to the user",
+              fullyQualifiedNamespace: "fully qualified namespace",
+              entityPath: "entity path",
+              errorSource: "renewLock"
+            }
+          );
+
           throw new Error("Whoops!");
         }
       },
@@ -29,7 +44,12 @@ describe("shared", () => {
       } as ServiceBusLogger
     );
 
-    wrappedProcessError(new Error("Actual error that was passed in from service bus to the user"));
+    wrappedProcessError({
+      error: new Error("Actual error that was passed in from service bus to the user"),
+      entityPath: "entity path",
+      errorSource: "renewLock",
+      fullyQualifiedNamespace: "fully qualified namespace"
+    });
 
     assert.isTrue(logErrorCalled, "log error should have been called");
   });