+++ sdk/servicebus/service-bus/test/smoketest.spec.ts
@@ -8,6 +8,7 @@ import chaiAsPromised from "chai-as-promised";
 import { getEntityNameFromConnectionString } from "../src/constructorHelpers";
 import { ServiceBusClientForTests, createServiceBusClientForTests } from "./utils/testutils2";
 import { ServiceBusSender } from "../src/sender";
+import { ProcessErrorArgs } from "../src/models";
 chai.use(chaiAsPromised);
 const assert = chai.assert;
 
@@ -56,8 +57,8 @@ describe("Sample scenarios for track 2", () => {
           await receiver.completeMessage(message);
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });
 
@@ -128,8 +129,8 @@ describe("Sample scenarios for track 2", () => {
         async processMessage(message: ServiceBusReceivedMessage): Promise<void> {
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });
 
@@ -211,8 +212,8 @@ describe("Sample scenarios for track 2", () => {
           await receiver.completeMessage(message);
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });
 
@@ -235,8 +236,8 @@ describe("Sample scenarios for track 2", () => {
         async processMessage(message: ServiceBusReceivedMessage): Promise<void> {
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });
 
@@ -348,8 +349,8 @@ describe("Sample scenarios for track 2", () => {
         async processMessage(message: ServiceBusReceivedMessage): Promise<void> {
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });
 
@@ -384,8 +385,8 @@ describe("Sample scenarios for track 2", () => {
         async processMessage(message: ServiceBusReceivedMessage): Promise<void> {
           receivedBodies.push(message.body);
         },
-        async processError(err: Error): Promise<void> {
-          errors.push(err.message);
+        async processError(args: ProcessErrorArgs): Promise<void> {
+          errors.push(args.error.message);
         }
       });