+++ sdk/storage/storage-file-datalake/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 12.2.0-beta.2 (Unreleased)
 
+- Bug fix - Fixes an issue where `DataLakePathClient.move()` will give an `InvalidSourceUri` error when the copy source name contains characters that need to be URL encoded. Fixed bug [11849](https://github.com/Azure/azure-sdk-for-js/issues/11849).
 
 ## 12.2.0-beta.1 (2020-10-13)